[Fix] 그라데이션 색상 구현 및 적용
