+++ PARD/Sources/+Font.swift
@@ -36,21 +36,3 @@ extension UIFont {
         let caption2 = UIFont.systemFont(ofSize: 12, weight: .semibold).withLineHeight(16)
     }
 }
-
-
-extension UIColor {
-    static let gradientColro = UIColor(patternImage: gradientImage())
-    private static func gradientImage() -> UIImage {
-        let gradientLayer = CAGradientLayer()
-        gradientLayer.frame = CGRect(x: 0, y: 0, width: 1, height: 1)
-        gradientLayer.colors = [UIColor(red: 82/255, green: 98/255, blue: 245/255, alpha: 1).cgColor, UIColor(red: 123/255, green: 63/255, blue: 239/255, alpha: 1).cgColor]
-        gradientLayer.startPoint = CGPoint(x: 0, y: 0)
-        gradientLayer.endPoint = CGPoint(x: 1, y: 1)
-        
-        UIGraphicsBeginImageContext(gradientLayer.bounds.size)
-        gradientLayer.render(in: UIGraphicsGetCurrentContext()!)
-        let image = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        return image!
-    }
-}
