+++ PARD/Sources/Home/HomeUpcommingView.swift
@@ -30,7 +30,7 @@ class HomeUpcommingView : UIView {
         $0.textAlignment = .center
         $0.text = "전체"
         $0.textColor = .pard.white100
-        $0.backgroundColor = .pard.gra
+        $0.backgroundColor = .GradientColor.gra
         $0.layer.cornerRadius = 4
         $0.layer.masksToBounds = true
     }
