+++ PARD/Sources/MyScore/ViewControllers/MyScoreViewController.swift
@@ -4,7 +4,6 @@
 //
 //  Created by 김민섭 on 3/4/24.
 //
-
 import UIKit
 import PARD_DesignSystem
 
@@ -760,4 +759,3 @@ extension MyScoreViewController {
         }
     }
 }
-
